Added scripts for preprocessing and building hdf5
The script for preprocessing takes an argument for the raw data directory, and the preprocessed data directory.
It reads through all files present in the raw directory, and applies average referencing, bandpass filters, notch filters, and ICA to clean artifacts.
The preprocessed file is store in the preprocessed data directory.

The script to build the comprehensive hdf5 dataset includes reading the preprocessed files, epoching them into fixed time lengths or on the evoked events, and then saving the concatenated data to a key within the hdf5 file for easy access.
Similarly, it creates shuffled reorders for different train/test splits.